UNI-97: decompose notes
