+++ package.json
@@ -85,7 +85,7 @@
 		"react-hook-form": "^7.53.2",
 		"react-router": "7.0.2",
 		"react-router-dom": "6.22.1",
-		"swr": "^2.2.5",
+		"swr": "2.2.5",
 		"yup": "^1.6.1",
 		"zustand": "^5.0.2"
 	},
