+++ src/entities/Note/index.ts
@@ -1,25 +1,9 @@
-import { Item } from './ui/Item/Item';
-import { List } from './ui/List/List';
-import { ListItem } from './ui/ListItem/ListItem';
-
-export type { INote, TNodeData, TNoteFormFields } from './model/types/note';
-
-type TNoteComponents = {
-	List: typeof List;
-	ListItem: typeof ListItem;
-	Item: typeof Item;
-};
-
-/** Экспортируем обертку для компонентов, которые связаны логически */
-export const Note: TNoteComponents = {
-	List,
-	Item,
-	ListItem,
-};
-
+/** Internal */
 export { useNoteStore } from './model/slice/useNoteStore';
 
-export { useGetNotes } from './api/useGetNotes';
-export { useCreateNote } from './api/useCreateNote';
-export { useUpdateNote } from './api/useUpdateNote';
-export { useDeleteNote } from './api/useDeleteNote';
+/** External */
+export { useGetNotes } from '@uniteam31/uni-shared-toolkit';
+//
+export { INote, TNodeData, TNoteFormFields } from '@uniteam31/uni-shared-types';
+//
+export { Note } from '@uniteam31/uni-shared-ui';
