--- src/entities/Note/ui/List/List.module.scss
@@ -1,22 +1,0 @@
-.List {
-    padding: 20px;
-    width: 321px;
-
-    .item {
-        &:hover {
-            // TODO в переменную
-            background-color: #EAEAEC;
-        }
-    }
-
-    .selected {
-        background-color: #EAEAEC;
-    }
-
-    .skeleton {
-        border-radius: 12px;
-        height: 80px;
-        margin-bottom: 5px;
-        width: 100%;
-    }
-}
