--- src/entities/Note/ui/TextArea/TextArea.module.scss
@@ -1,12 +1,0 @@
-.TextArea {
-    all: unset;
-
-    //
-    background-color: #fff;
-    font: var(--font-body);
-    font-weight: 300;
-    height: 100%;
-    overflow-wrap: break-word;
-    resize: none;
-    width: 100%;
-}
