--- src/entities/Note/ui/TitleInput/TitleInput.module.scss
@@ -1,10 +1,0 @@
-.TitleInput {
-    all: unset;
-    background-color: #fff;
-
-    //
-    font: var(--font-subheadline);
-    font-weight: bold;
-    height: auto;
-    width: 100%;
-}
