--- src/pages/NotesPage/index.ts
@@ -1,1 +1,0 @@
-export { NotesPageAsync as NotesPage } from './ui/NotesPage/NotesPage.async';
