--- src/pages/NotesPage/ui/Form/Form.module.scss
@@ -1,5 +1,0 @@
-.loader {
-	position: relative;
-	height: 100%;
-	flex: 1;
-}
