--- src/pages/NotesPage/ui/NotesPage/NotesPage.async.tsx
@@ -1,4 +1,0 @@
-import { lazy } from 'react';
-
-
-export const NotesPageAsync = lazy(() => import('./NotesPage'));
