+++ src/shared/api/index.ts
@@ -1,0 +1,1 @@
+export { axiosInstance } from '@uniteam31/uni-shared-toolkit';
