--- src/shared/hooks/useDebounceValue/useDebounceValue.ts
@@ -1,18 +1,0 @@
-import { useState, useEffect } from 'react';
-
-// eslint-disable-next-line
-export const useDebounceValue = (value: any, delay: any) => {
-	const [debouncedValue, setDebouncedValue] = useState(value);
-
-	useEffect(() => {
-		const handler = setTimeout(() => {
-			setDebouncedValue(value);
-		}, delay);
-
-		return () => {
-			clearTimeout(handler);
-		};
-	}, [value, delay]);
-
-	return debouncedValue;
-};
