+++ src/shared/types/index.ts
@@ -1,0 +1,1 @@
+export type { TMeta, ApiResponse, TAccessToken } from '@uniteam31/uni-shared-types';
