--- src/shared/types/meta.ts
@@ -1,6 +1,0 @@
-/** Мета данные, которые приходят с каждым запросом сущности с сервера */
-export type TMeta = {
-	_id: number;
-	createdAt: string;
-	updatedAt: string;
-};
